test(echo_service): refactor the echo_service to be a unit test instead of integration
