+++ src/tests/mod.rs
@@ -1,2 +1,2 @@
 mod common;
-mod igd;
+mod echo_service;
